Remove observer parse from api (#816)
