+++ pkg/blockatlas/observer.go
@@ -1,8 +1,12 @@
 package blockatlas
 
+import "strconv"
+
+type Subscriptions map[string][]string
+
 type Webhook struct {
-	Subscriptions map[string][]string `json:"subscriptions"`
-	Webhook       string              `json:"webhook"`
+	Subscriptions Subscriptions `json:"subscriptions"`
+	Webhook       string        `json:"webhook"`
 }
 
 type CoinStatus struct {
@@ -13,4 +17,22 @@ type CoinStatus struct {
 type Observer struct {
 	Status  bool   `json:"status"`
 	Message string `json:"message"`
+}
+
+func (w *Webhook) ParseSubscriptions() []Subscription {
+	subs := make([]Subscription, 0)
+	for coinStr, perCoin := range w.Subscriptions {
+		coin, err := strconv.Atoi(coinStr)
+		if err != nil {
+			continue
+		}
+		for _, addr := range perCoin {
+			subs = append(subs, Subscription{
+				Coin:    uint(coin),
+				Address: addr,
+				Webhook: w.Webhook,
+			})
+		}
+	}
+	return subs
 }
